[Feat/#114] DayRecordsManager 로 분리
